requirements.txt
APScheduler==3.10.4
backports.zoneinfo==0.2.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.0.4
xmltodict==0.13.0
zipp==3.20.1
Flask-Migrate==4.0.7
openai==0.28
scikit-learn==1.3.2

